COL-1775, consolidate 'init' phase and simplify vuex mutation phase
