+++ .env.development
@@ -3,5 +3,7 @@
 # Only variables that start with 'VUE_APP_' will be statically embedded into the client bundle.
 ####
 
+FABRIC_JS_DEBUG_EVENTS_EXCLUDE = ['event:mouseout', 'event:mouseover']
+
 VUE_APP_API_BASE_URL='http://localhost:5000'
 VUE_APP_DEBUG=true
