+++ src/store/whiteboarding/mutations.ts
@@ -1,11 +1,10 @@
 import _ from 'lodash'
-import canvas from '@/store/whiteboarding/utils/canvas'
 import FABRIC_MULTIPLE_SELECT_TYPE from '@/store/whiteboarding/utils/constants'
 import fabricator from '@/store/whiteboarding/utils/fabricator'
 import stateDefault from '@/store/whiteboarding/utils/state-default'
 import utils from '@/api/api-utils'
 import Vue from 'vue'
-import socket from '@/store/whiteboarding/utils/socket'
+import {initialize} from '@/store/whiteboarding/utils/initializer'
 
 const p = Vue.prototype
 
@@ -78,15 +77,7 @@ export default {
       sidebarExpanded: !whiteboard.deletedAt,
       viewport: document.getElementById('whiteboard-viewport')
     })
-    canvas.init(state)
-    // listeners.init(state)
-    if (!whiteboard.deletedAt) {
-      // Open a websocket connection for real-time communication with the server (chat + whiteboard changes) when
-      // the whiteboard is rendered in edit mode. The course ID and API domain are passed in as handshake query parameters
-      socket.init(state, whiteboard)
-    }
-    // The whiteboard p.$canvas should be initialized only after our additions are made to Fabric prototypes.
-    fabricator.init(state)
+    initialize(state, whiteboard)
   },
   moveLayer: (state: any, direction: string) => {
     /**
